% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{char_to_num}
\alias{char_to_num}
\title{char_to_num}
\usage{
char_to_num(char)
}
\arguments{
\item{char}{a character string}
}
\value{
a vector with numbers which correspond to the character string
}
\description{
char_to_num
}
