% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{crop_func}
\alias{crop_func}
\title{crop_func}
\usage{
crop_func(page, crop_area, text_shape, median_filter_size, pad)
}
\arguments{
\item{page}{magick pointer}

\item{crop_area}{a string with the crop geometry}

\item{text_shape}{shape of the output array}

\item{median_filter_size}{vector with radius for two median filters}

\item{pad}{the size of the padding around the crop area}
}
\value{
an array with the text crop
}
\description{
crop_func
}
