% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNN_model_optimized.R
\name{decode_ctc}
\alias{decode_ctc}
\title{decode_ctc}
\usage{
decode_ctc(model, images)
}
\arguments{
\item{model}{The Keras model pointer}

\item{images}{An array with cropped images of the LP and VIN}
}
\value{
A tibble with the coresponding decoded text and probability to each of the images.
}
\description{
decode_ctc
}
