% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{first_stage_pre_process}
\alias{first_stage_pre_process}
\title{first_stage_pre_process}
\usage{
first_stage_pre_process(pdf, origin_shape)
}
\arguments{
\item{pdf}{magick pointer}

\item{origin_shape}{the shape of the array}
}
\value{
an array with resized pages
}
\description{
first_stage_pre_process
}
