% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{format_text_output}
\alias{format_text_output}
\title{format_text_output}
\usage{
format_text_output(text_pred)
}
\arguments{
\item{text_pred}{a character string}
}
\value{
a cleaned character string
}
\description{
This function does apply the BR for the VIN text
}
