% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{hough_line_votes}
\alias{hough_line_votes}
\title{hough_line_votes}
\usage{
hough_line_votes(Z, degree, d_rho)
}
\arguments{
\item{Z}{a binary Canny filter image}

\item{degree}{vector with possible rotations}

\item{d_rho}{the resolution of rho}
}
\value{
an estimated rotation in degrees
}
\description{
hough_line_votes
}
