% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{iLD}
\alias{iLD}
\title{Iterative Levenshtein Distance}
\usage{
iLD(s, t)
}
\arguments{
\item{s}{string}

\item{t}{truth}
}
\value{
an iteger with the number of operations to perform i order to get the correct string
}
\description{
Iterative Levenshtein Distance
}
