% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{image_preprocess}
\alias{image_preprocess}
\title{image_preprocess}
\usage{
image_preprocess(pdf_file, page_idx = NULL)
}
\arguments{
\item{pdf_file}{the path of the the pdf file}

\item{page_idx}{default null, provide vector for specific pages in the VRD}
}
\value{
an updated magick pointer
}
\description{
image_preprocess
}
