% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNN_model_optimized.R
\name{model_CNN}
\alias{model_CNN}
\title{model_CNN}
\usage{
model_CNN(no_classes = 4, batch_shape = list(64, 224, 224, 3),
  lr = 0.001, beta_1 = 0.9, beta_2 = 0.999)
}
\arguments{
\item{no_classes}{an integer, number of classes.}

\item{batch_shape}{a list with `batch_shape`.}

\item{lr}{the learning rate.}

\item{beta_1}{a exponential decay, 1st moment estimates.}

\item{beta_2}{a exponential decay, 2nd moment estimates.}
}
\value{
Keras model object
}
\description{
A VVG16 model with a conv base which is pretrained on the ImageNet data and the self defined classifier.
}
