% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{my_pipe}
\alias{my_pipe}
\title{Title}
\usage{
my_pipe(pdf_file, page_idx = NULL)
}
\arguments{
\item{pdf_file}{the path to the VRD pdf file we wish the provide an classification and extract to.}

\item{page_idx}{default null, provide vector for specific pages in the VRD}
}
\value{
a list with lists
}
\description{
Title
}
