% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{OTSU}
\alias{OTSU}
\title{OTSU}
\usage{
OTSU(input)
}
\arguments{
\item{input}{a vector of gray-scale pixel values}
}
\value{
The OTSUs theshold
}
\description{
The function determine the Otsus threshold
}
