% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pipe.R
\name{second_stage_pre_process}
\alias{second_stage_pre_process}
\title{Title}
\usage{
second_stage_pre_process(page, text_shape, median_filter_size = c(5, 3),
  pad = 4)
}
\arguments{
\item{page}{magick pointer}

\item{text_shape}{output shape for text field}

\item{median_filter_size}{a vector of two, default c(5,3)}

\item{pad}{the padding, default 4}
}
\value{
a list with two array, one for LP and one for VIN
}
\description{
Title
}
