% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypipe-package.R
\docType{package}
\name{thesispipe-package}
\alias{thesispipe}
\alias{thesispipe-package}
\title{thesispipe: The R package which produces the end-to-end results of the technical pipeline on the test set}
\description{
This R package is the final result of my master thesis. 
             It have been created in order to produce the end-to-end results on the test set.
}
\author{
\strong{Maintainer}: Anders Bæk \email{anders@launer.dk}

}
\keyword{internal}
