% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vocab_tot}
\alias{vocab_tot}
\title{A character vector with valid tokens plus *}
\format{A vector}
\usage{
vocab_tot
}
\description{
A character vector with valid tokens plus *
}
\keyword{datasets}
