% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNN_model_optimized.R
\name{model_CNN}
\alias{model_CNN}
\title{model_CNN}
\usage{
model_CNN(no_classes, batch_shape, glo_set)
}
\arguments{
\item{no_classes}{an integer, number of classes.}

\item{batch_shape}{a list with `batch_shape`.}

\item{glo_set}{a list must be `glo_set`.}
}
\value{
Keras model object
}
\description{
VVG 16 pretrained on image net
}
