% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNN_model_optimized.R
\name{model_CNN_RNN}
\alias{model_CNN_RNN}
\title{model_CRNN
inspired by: https://github.com/keras-team/keras/blob/master/examples/image_ocr.py
and https://github.com/DeepSystems/supervisely-tutorials/blob/master/anpr_ocr/src/image_ocr.ipynb}
\usage{
model_CNN_RNN(vocab_size, batch_shape, max_str_len, lr, beta_1, beta_2)
}
\arguments{
\item{batch_shape}{a list with `batch_shape`.}

\item{no_classes}{an integer, number of classes.}

\item{glo_set}{a list must be `glo_set`.}
}
\value{
Keras model object
}
\description{
The CRNN model
}
